Fix filename check when creating output zip file
Fixes #76.

In some cases, the Gdal output file isn't just `${layerName}.${extension}`. (This problem originally came up when there was a period in the layer name, but there may be other such cases.)

Anyway, Gdal actually tells us the name of one of the files it output for the last ogr2ogr command, so use that to pick out the output files to zip up.